+++ jest.config.js
@@ -7,14 +7,15 @@ module.exports = {
   },
   testEnvironment: 'node',
   coverageDirectory: './coverage',
-  collectCoverageFrom: ['apps/**/*.(t|j)s'],
+  coverageReporters: ['text', 'lcov', 'json-summary'],
   projects: [
     {
       displayName: 'shop',
       rootDir: 'apps/shop',
+      setupFilesAfterEnv: ['<rootDir>/../../jest.setup.ts'],
       testMatch: ['<rootDir>/src/**/*.spec.ts'],
       transform: {
-        '^.+\\.(t|j)s$': 'ts-jest',
+        '^.+\.(t|j)s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
       },
       moduleNameMapper: {
         '^@/(.*)$': '<rootDir>/src/$1',
@@ -22,13 +23,23 @@ module.exports = {
         '^@app/common$': '<rootDir>/../../libs/common/src',
       },
       testEnvironment: 'node',
+      collectCoverageFrom: [
+        'src/**/*.(t|j)s',
+        '!src/**/*.spec.ts',
+        '!src/**/*.module.ts',
+        '!src/main.ts',
+        '!src/data-source.ts',
+        '!src/proto/**',
+        '!src/db/migrations/**',
+      ],
     },
     {
       displayName: 'payments',
       rootDir: 'apps/payments',
+      setupFilesAfterEnv: ['<rootDir>/../../jest.setup.ts'],
       testMatch: ['<rootDir>/src/**/*.spec.ts'],
       transform: {
-        '^.+\\.(t|j)s$': 'ts-jest',
+        '^.+\.(t|j)s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
       },
       moduleNameMapper: {
         '^@/(.*)$': '<rootDir>/src/$1',
@@ -36,6 +47,15 @@ module.exports = {
         '^@app/common$': '<rootDir>/../../libs/common/src',
       },
       testEnvironment: 'node',
+      collectCoverageFrom: [
+        'src/**/*.(t|j)s',
+        '!src/**/*.spec.ts',
+        '!src/**/*.module.ts',
+        '!src/main.ts',
+        '!src/data-source.ts',
+        '!src/proto/**',
+        '!src/db/migrations/**',
+      ],
     },
   ],
 };